Changes EventStream end behavior to not reconnect
Now when the response emits an end event, the EventStream will also emit
an 'end' event and then remove all of its listeners. It will be up to
the application programmer to decide what to do in that event. The
library will not automatically try to reconnect.

- Changes assignments that were using `undefined` to `null`. Null should
indicate intentional emptyness, while undefined means it truly was never
defined.

- Makes other small formatting changes